Fix issues and implement calendar feature
